card-downloader/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==11.2.1
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
rich-pixels==3.0.1
tqdm==4.67.1
urllib3==2.4.0
zstandard==0.23.0
